将DBHelper的代码进行拆分，提高代码整洁度:zap:
