+++ 10-Code/SevenTiny.Bantina.Bankinate/DataAccessEngine/DataBaseType.cs
@@ -1,0 +1,24 @@
+﻿/*********************************************************
+* CopyRight: 7TINY CODE BUILDER. 
+* Version: 5.0.0
+* Author: 7tiny
+* Address: Earth
+* Create: 1/7/2019, 3:42:48 PM
+* Modify: 
+* E-mail: dong@7tiny.com | sevenTiny@foxmail.com 
+* GitHub: https://github.com/sevenTiny 
+* Personal web site: http://www.7tiny.com 
+* Technical WebSit: http://www.cnblogs.com/7tiny/ 
+* Description: 
+* Thx , Best Regards ~
+*********************************************************/
+namespace SevenTiny.Bantina.Bankinate.DataAccessEngine
+{
+    public enum DataBaseType
+    {
+        SqlServer,
+        MySql,
+        Oracle,
+        MongoDB
+    }
+}
